glotlib: Add append_x_y_data() to append points in amortized O(1) time.
Previously, the only way to extend a series was to use one of the
set_x_y_data()-family methods.  This would have to write the entire data
set back to the VBO.  This commit adds append_x_y_data() methods that
allow writing just the new entries to the VBO, extending the data buffer
on the GPU if necessary.  The extensions increase by powers-of-2 so that
the append operation is amortized O(1).